requirements.txt
fastapi
uvicorn
python-telegram-bot==20.0
openai